Reimplement automatic call of `Mkspell()`
When opening a file that does not exist on startup, only the last line
of `mkspell`'s message, which is not informative, is displayed. To
always print the entire message, call `Mkspell()` from `vimrc` (only)
upon initialization, and let `Mkspell()` define an autocommand that will
subsequently call `Mkspell()` at every `BufWinEnter` event.