Re-position custom VimTeX completion files
Instead of placing a directory with a copy of all of the default VimTeX
completion files plus my own completion files, and putting that
directory at the tip of my `runtimepath` so that it is loaded instead of
the normal VimTeX completion file directory, maintain a separate
repository just of my own completion files, include it as a submodule of
my home repository, and clone it into VimTeX's completion file
directory. That way, I can track changes to my completion files,
maintain them in a separate place from VimTeX's completion files so that
they are not lost if e.g. I uninstall VimTeX, not mess with my
`runtimepath`, and not have a hoard of duplicate files.