Fix TeX indentation bug in surround mappings
Because surround mappings begin with closing delimiters, under certain
circumstances (e.g., when in an list environment with items) executing
the mapping causes the line to indent backwards. This causes the visual
mode mappings, which rely on the (fixed) location of the ``<` and ``>`
marks, to not function properly. Lead with the opening delimiter to fix
this problem. (Also fix some indentation and add an `iabbrev`).