Rewrite check for previously defined aliases
Earlier, checked whether or not alias definition matched element in
output of `alias` command; now, simply check whether or not alias name
was used earlier in `zshrc`. This way, when `whence` finds an alias, an
error is only thrown if the user is attempting to use the same alias
name twice. That way, no error messages will be thrown when e.g.
`.zshrc` is re-sourced