Load `geometry` with margins dependent on font size
Only load when standard LaTeX class is used, in one-sided and one-column
mode, and if letter paper is used. The horizontal margins are based on
the default horizontal margins from the LaTeX standard `article` class;
`vmarginratio` is set in order to make the text look centered (if the
text is actually centered -- i.e., if the top and bottom margins are
equal -- the text looks like it is sagging. Therefore, add a little
extra to the bottom margin)