Add YAML headers to index files
Add default header to new index files, and use existing header for
existing index files. Because it's simplest to just append links to the
index file, we need a way to avoid duplicating all the links every time
the script is run. The simplest way to do this seems to be to just
create a new index file every time the script is run. Therefore, add a
way to get the YAML header from any existing index file, so that we can
add that same header to the new index file.

*   So that we can use keywords in index files without adding links to
    index files to index files, only scan notes files for keywords when
    building the index
*   Use underscores instead of spaces in index file names; use this
    filename (sans extension) as the id of the index note