Add outline formatting to index files
For every index file corresponding to a given keyword, make a section in
that index file for every other index file that has the given keyword as
one of its keywords. In each of these sections, put the notes that have
both keywords corresponding to each index file (not including index
files). This makes it easy to see when looking at the 'De anima' index
file, for example, which notes are both about e.g. phantasia and the De
anima. Sort the sections in alphabetical order, and sort other notes by
ID (which turns out to be chronologically).