Add advanced support for additional `ripgrep` or `fzf` queries
Press ctrl-e to filter the results again with a new ripgrep query, and
press ctrl-w to filter the results again with a new fzf query. Can be
done an arbitrary number of times, in any order. Accomplished by storing
results in temporary files.

*Known issue*: there may be a discrepancy between what `fzf` finds in
the interactive interface and what `fzf --filter` finds due to matching
of ANSI color codes (this could be rectified by using `--ansi`, but
this also eliminates colors from the output. In practice, it's better to
accept this issue).