`rfv`: consolidate array elements
There's no reason to have so many array elements in the `fzf` and `rg`
expressions, because there's no need to control word splitting, because
`fzf` feeds the commands in `reload()` and `execute()` to `zsh -c` as
one word (i.e., as if `zsh -c "$stuff_in_parentheses"`). So just split
strings into array elements as needed, not at every word.