`nvimrc`: Improve management of global vs filetype options
For all options that are set both globally and as part of a `FileType`
autocommand, only set the global options once. That is, once `init.vim`
has been sourced, don't set the global options again if ever `init.vim`
is re-sourced. Accomplish this by setting a switch at the end of
`init.vim`, and setting the global options in question on the condition
that the switch is off. This prevents annoying things like text files,
which according to a `FileType` command do not have `number` set, all of
a sudden being numbered when `init.vim` is re-sourced.