`build-index`: Support multiple YAML headers
Enclose parts of `_get_block` in additional `while` loop that runs until
there are no elements left in `line_numbers`. Along the way, run all the
same checks that we did to detect a single YAML block; but, because
we're running through all the elements in `line_numbers`, we can detect
multiple YAML blocks. Cache their contents as one string in associative
array `blocks`. When parsing blocks, use the `I` reverse-indexing flag
instead of the `i` reverse-indexing flag to look up the line with the
key, in order to find the value provided by the last of the YAML blocks,
like Pandoc does.

Bug fix: change `while true` loop to `while (( $#line_numbers > 1 ))` to
ensure that enough arguments are passed to `__check_lines`