`build-index`: Make `_get_block` perfectly general and precise
Use PCRE2 regular expressions to detect all and only YAML keys and
thereby detect all and only YAML blocks with great precision. However,
this slows down build-index considerably (because each call of
_get_block requires multiple invocations of pcre2grep), so, as beautiful
as it is, this functionality should be abandoned. For the purposes of
build-index, it will suffice to ensure that 'title:' and 'keyword:' are
contained within the header; it is not necessary to detect other keys