Simplify `alias` and `autoload` functions
Whenever the either function is called in .zshrc, check whether or not
the arguments are in an array, where the array contains
functions/aliases that were defined previously in .zshrc. If the alias
or function was previously defined but is in that array, don't throw an
error. Otherwise, throw an error. Compared to the previous
implementation, this has the disadvantage that the function won't throw
an error for anything multiply defined in the .zshrc; but the user
should be able to catch this. Otherwise, the code is simpler, and
nothing additional is required in the .zshrc for the functions to work.

For `autoload`, additional advantage is that, since autoloaded functions
in `.zshrc` are `unfunction`ed one at a time instead of in a group,
no errors are thrown for functions autoloaded by `compinit` but for some
reason no longer defined by the time `.zshrc` is reloaded