Refactor `notes#exit_note`
Instead of running Mdviewconvert and build-index at BufWinLeave, define
autocommands for BufEnter or VimLeavePre (depending on b:exiting) when
BufWinLeave occurs. This is useful when exiting because then a whole
bunch of VimLeavePre autocommands (one for each note buffer edited
during a session) don't proliferate. This is useful when not exiting
because then it's not necessary to force a 'hit-enter' prompt