Rewrite `notes#exit_note`
Use 'BufWinLeave' event directly, not to set 'BufEnter' event (the
initial idea behind this strategy was that it made manually programming
a 'hit enter' prompt unnecessary, but it caused many other problems).
Use scratch buffer for mdview_convert instead of echoing so that output
is displayed even if using rfv.