Partially rewrite function for renumbering lists
Use list objects to identify previous and next items instead of
navigation, which can be faulty (especially with child items).

In the future, probably compare index of current item object to the
index of the script-local item object saved when calling
s:change_numbered_list. Use the offset to determine by how much to
change the labels, rather than the difference between the label of the
current list item and the label of the next list item.