Rewrite inner and outer item object code
Use same core function for both, and utilize parent-finding code
originally in outer list item code to get information about parent
objects for both inner and outer item objects. After the item has been
acquired, modify it to make it meet the specifications of an inner or
outer item object:
*   If it is an inner object, set the ending line before any child
    items, unless there is text following the child items
*   If it is an outer object, reset just about everything to accord with
    the indent and so on of the first line of the outer object
Get the core list item (i.e., the numbered or bulleted line, followed by
everything indented in accordance with it) with one function, and do the
foregoing with another function. A function is also devoted to finding
the end of an inner item object. Change the code throughout to accord
with the changes in nomenclature.