Consolidate outer and inner list object functions into one
Whenever a:item is passed to an argument, make a deep copy to avoid
inadvertendtly changing it (when necessary). In the same spirit, with
get_children and get_parents, don't modify list directly -- just return
list of children or parents.

Modify get_item so that a line number can be passed to it as an
argument.

Shorten and clarify some variable names.

Tested, and navigation still works after replacing all the old function
references.