Consolidate code common to both list and item navigation
Use helper functions for navigating within lists and for navigating
between lists. Use one function for gathering possible jump destinations
for items and another for lists; there are too many differences between
items and lists for there to be a common function for this.

Write code so that jumplist is updated.

Instead of passing an item to list_object, pass a list; this is
symmetrical with item_object, to which one passes an item.