Find documents in Zotero storage by parsing BibLaTeX file
Construct an array of biblatex keys and an array of the line numbers at
which those keys are located. Search for files in the Zotero cloud
storage directory: construct an array of those files and another array
of the line numbers at which those files are located. For each line
number in the array of filepath line numbers, find the line number of
the biblatex keys that's closest before it. Then we know that the
biblatex key at that line number is the biblatex key that's assorted
with the attachment, and we can copy the attachment to the appropriate
directory.

Another significant change: For each attachment that is copied, ask what
subdirectory to put it in. I.e., it is now required that every document
be located in a subdirectory. This makes clear what the document is
(clean, annotated, etc.) always, and it prevents any possible naming
conflicts/clobbering. Accordingly, don't make 'clean' and 'annotated'
subdirectories by default, and don't clean up empty subdirectories at
the start of the script (there shouldn't be any).

Make various changes to variable names.