Add quickfix functionality, and define `s:print_msg`
Consolidate code for printing messages into s:print_msg, and use that in
callback functions to print messages, if 'msg' is enabled.

Add option to parse output and create a quickfix list. Option to open a
window or not, set the window's title, and set the window's height. If
needed, someone could define a custom function in place of the default
function.