Simplify opt structure and rewrite code as object
Because the code was based heavily around jobstart's 'opts' dictionary
anyway.

Instead of 'msg'=0 or 1 and 'display_stdout' and 'display_stderr', just
use msg=0 through 3 (0 for none, 1 for stdout, 2 for stderr, 3 for
both).

Instead of 'qf'=0 or 1 and 'qf_args', just set 'qf' equal to an opts
dictionary, and only run 'createqflist' if key is present

Because the object is defined before 'shell#jobstart' is executed,
callback functions can be overwritten with custom functions passed to
'shell#jobstart' (the original values will be overwritten when 'extend'
is called)