Write separate `compiler` object and function
To handle jobs with quickfix lists. Instead of just accepting one
callback function, accept a list of callback functions, and insert the
function for creating the quickfix list at the beginning of the list.
This adds no new functionality, but quickfix lists and shell commands
per se seem pretty orthogonal, so it seemed strange to have
'shell#jobstart', which manages shell commands in general, handle
quickfix lists. It made sense to have a separate object and function to
handle that subset of shell commands that would call for a quickfix
list (i.e., compilers)