Improve comment mappings and option setting
Improve comment mappings by using functions to edit the lines that are
to be commented or un-commented instead of normal-mode commands. This
allows us to use '<Cmd>' instead of ':' in the mappings, and it allows
us to do different things depending on whether or not the line is empty
(so no more annoying trailing spaces if an empty line is commented).

Improve option setting by, instead of only setting global options that
conflict with local options on startup, using setfiletype to reload
filetype-specific local options after the global options have been set.
This means that, as one might expect, all options will always be
applied, and there's no need to use the 's:sourced' variable to keep
track of whether or not 'init.vim' is being sourced on startup