Simplify logical structure of code and add `-c` opt
Use '-c' option in 'autoload' and 'alias' to mark functions and aliases,
respectively, that are in '.zshrc' and should not be marked as 'already
defined' if '.zshrc' is re-sourced. The old approach was to say that any
function or alias in '.zshrc' should not be marked as 'already defined'.
This way is better, because it allows us to distinguish between e.g.
user-defined functions and e.g. 'compinit' in '.zshrc', such that
user-defined functions are not marked as 'already defined', but
'compinit' is marked as 'already defined' when '.zshrc' is re-sourced
(which allows us to execute or not execute certain commands depending on
whether or not 'compinit' has been marked as 'already defined').