Add support for loading output into scratch buffer
Use name of command to create buffer-local dictionary containing the
attributes of the scratch buffer (the bufnr, so that the bufwinid can be
obtained and so that we can print text to the buffer without switching
windows; and a 'completed' flag, set on exit, so that we can know if the
command has executed and if we need to reuse the current scratch
buffer).

Use 'scratch_win' key to configure title and height of scratch window.
Create the buffer-local scratch buffer object as a deepcopy of
'scratch_win'. If a title is already used by a *listed* buffer, use a
title with the current buffer name appended in parentheses. On exit,
provide the option to choose one title if exit status is zero, and
another title if the exit status is non-zero.

Allow 7 options for 'msg' or for 'scratch': 0, 1, 2, 3, 4, 5, or 6. 0-3
mean the same as they did; 4-6 mean the same as 1-3, but 'print_msg' and
'set_scratch_buffer' will only execute on exit. Executing on exit should
be mandatory for e.g. 'cmdheight' to be reset and for the 'completed'
flag to be set, so an option is not provided to execute only on stderr
or only on stdout.

Code is currently more complicated than it needs to be because it allows
for more than one scratch buffer to be open at a time. In the future,
simplify to not allow this; after all, there is only one quickfix window
allowed at a time, and only one message window.