Reimplement error handling in `shell#compile`
Instead of making 'createqflist' the default callback function, write a
separate error handler that's called directly by the exit handler. If
'errorformat' is non-empty, the error handler will call 'createqflist',
and if 'errorformat' is empty, the error handler will print stderr to a
scratch buffer.

The idea is that, when commands are executed asynchronously and we want
to display the output, it makes sense to load output into a buffer of
some sort, as opposed to printing it to the message window, where it
might surprise the user and might be overwritten by the user. Using a
quickfix list is best, but some commands (e.g. Pandoc) don't produce
error messages of a consistent output that can be parsed using
'errorformat'. These, then, should just be printed verbatim to a scratch
buffer in lieu of a quickfix window.