Revise output handling and add scratch buf highlighting
Store output as list of dictionaries, each with a 'line' value and an
'event' value. That way, it can be established non-heuristically what
stream each line of output came from. Use 'both', 'stdout', and 'stderr'
functions to easily get a list of lines using 'map' and 'filter'.

Use the information encoded in 'output' to determine which lines in a
sceratch buffer are to be highlighted.

Small changes:

*   Revise comments in 's:on_output'
*   Make 'bufwinid' an attribute of 's:scratch_win'
*   Set 'l:scratch' and 'l:msg' equal to a *copy* of 'self.scratch' and
    'self.msg', respectively