Fix bug where scratch buffer not created in time
'getmatches' is called with 's:scratch.bufnr' as its argument, and then
'loadscratchbuf' is called. This means that, if 'Mdview' doesn't produce
any errors, it might be the case that no scratch buffer exists when
'getmatches' is called, which means 's:scratch' is empty and Vim will
complain that 'bufnr' is not a key of 's:scratch'. Call 'getmatches'
earlier, on the condition that 's:scratch' is non-empty.

Also note another bug with a 'FIXME' at the top of the file.