Use `scratch_buf` rather than `scratch_win` object
To better account for possibility that scratch window has closed, and to
make sure that we are only using one buffer. Check for mdView scratch
buffer highlighting depending on whether or not the window is open; if
it is not open, set matches to empty. Set the matches as part of the
'mdview' dictionary.

Also improve title handling. Reset the title when mdView is called on
exit. If there is no mdView output, reset title to a title appropriate
to build-index, if build-index has output. Append build-index title to
current buffer title only if there has also been mdView output.