Use nested key/value sets to easily pass `biblatex` options
The key 'authordate' or the key 'notes' is required. Upon detecting one
of these, store the key name in a macro. Then, depending on the
definition of the macro, execute another macro that loads
'biblatex-chicago' and executes default options. Any other options
passed to 'rhelder.sty''s 'biblatex' key are then executed after the
defaults, which allows the user to override the defaults if desired.

Also reorganize the package, again.