Change variables in light of changes to 'html' template
Use 'rhandout.class' instead of 'class', since 'class' is now going to
be used to encode course names, instead of just including the course
name in the template (that way, the metadata is specified in the source
files, and they can (mostly) reliably be recompiled if needed), and in
any even 'class' didn't do anything independently of 'rhandout' anyway.

Also include default author, since, again, we are no longer going to
include that in the template.