Make `\parsep` in lists same as `\parskip` in document
The goal is and was to make paragraph breaking the same within lists as
it is outside of lists. '\parsep' was just set to 0pt, which is fine if
e.g. 'parskip' package is not used. But now sometimes I do use the
'\parskip' package (e.g. in the 'rheldersyllabus' document class), and
in any event it is more precise to set 'parsep' not to 0pt but to
whatever 'parskip' is in the surrounding document.