Delete unnecessary Pandoc config files
Delete 'Makefile' and instead use a shell script to generate a Makefile
suitable for the current workflow. Delete 'handout.yaml' and instead
specify metadata on a project-by-project basis, so that it is more
transparent what metadata the project is actually using, especially if
somewhere down the line there is a change to the centralized metadata
file. Accordingly, from now on, only use centralized defaults and
metadata files for settings that are meant to apply to almost every file
generated by Pandoc.