Move `rhelder.sty` and change to `rhelder.dtx`
Move to a separate development directory outside of the texmf tree.
Instead, we will install 'rhelder.sty' in the texmf tree using
'l3build'. This way, we keep the source code and the files generated
from the source code separate.

To that end, switch to the 'dtx' file format so we can document
'rhelder.sty''s code alongside the code itself: 'rhelder.sty''s code has
gotten more complex, so this is now a useful thing to do.