Add first draft of `dtx` module
Toggle between a 'standard' mode (in which the `dtx` file is displayed
with its standard syntax) and an 'editing' mode (in which the `dtx` file
is more or less displayed as `doc` sees it -- i.e., with the comment
characters at the start of a line removed. This allows one to take
advantage of VimTeX's usual functionality (a lot of which does not work
if there is a comment character at the start of the line).

Also:
*   Disable some more unwanted folds
*   Delete some old and unused options and mappings