Improve LuaSnip configuration and unset `smarttab`
'smarttab' (NeoVim default) was causing 'shiftwidth' and 'softtabstop'
to not behave properly.

An unfortunate quirk of LuaSnip's is that, if you leave the snippet
without jumping all the way through the snippet (e.g., by switching to
Normal mode and then hitting 'o' to edit the next time), hitting <Tab>
will send you back to the snippet. This isn't an issue that the
developer has resolved satisfactorily (IMHO), but a user did post a
helpful bit of code in an issue discussing this quirk. Include this to
improve the LuaSnip experience.