Fix bug in LuaSnip config
When exiting insert mode, including for insert mode completion, current
snippet was removed from the jumplist. This meant that the snippet was
removed from the jumplist every time e.g. autocompletion activated.
Ensure that the snippet is only removed from the jumplist if we're
leaving insert mode *and* we're not leaving insert mode for completion.

Also change 'update_events' key in LuaSnip 'setup' so that replacement
nodes are updated every time text is entered.