Bug fix: change `guicursor` setting
When 'guicursor' was set to '', the cursor would disappear after closing
a popup menu. This is a bug in Neovim, but also kind of makes sense --
Neovim is generally using GUI colors, so it makes some sense that using
the cursor with just the terminal settings would cause problems in edge
cases. In any event, setting the cursor to 'block' in all modes seems to
have solved the problem for now.