Switch over to using `vim-pandoc`
In notes files, continue using same completion config as before, but
otherwise make use of `vim-pandoc` completion -- which is not
necessarily better, but it's good to not re-invent the wheel, in
principle.

In general, reasons to switch to `vim-pandoc`:
*   As mentioned, to not re-invent the wheel (including completion,
    links, and asynchronous execution of Pandoc, all of which I have my
    own code for)
*   Syntax highlighting
*   Motion commands

That said, the way that the plugin handles stderr from `pandoc` annoys
me. There's not nothing to be said for reinventing the wheel.