fix: match ordered list items correctly
The regex in the 'pandoc' syntax file included with Neovim does not
require white space between the list marker (e.g. '1.') and the list
item text, which leads to a false match when the line begins with e.g.
'6.7, 1141b5-7', in which case the '6' at the beginning of the line is
highlighted.

Also, now that the deprecated 'vim-pandoc-syntax' has been uninstalled,
set 'g:vim_pandoc_syntax_exists' to '1', so that 'vim-pandoc' knows that
'pandoc' syntax highlighting is available.