fix: rewrite broken `shell#output_to_quickfix`
Initialize the quickfix list. If the current quickfix list is associated
from the job (using the quickfix list's title as a heuristic), reset the
list. Otherwise, create a new quickfix list. The *add* the results of
evaluating an expression to the quickfix list, and then set the title.
Use `botright cwindow` to make sure that the quickfix window occupies
the full width of the screen. `cwindow` closes the quickfix window if
there are no errors, so we don't need to check for that.

Simplify title handling for both scratch buffers and quickfix lists
using the utility function 's:title()'. If the argument to 's:title()'
is empty, then return a default title based on the command passed to
'jobstart()'.