feat!: support multiple scratch buffers
Using the script-local 's:scratch_buf' to store data about the scratch
buffer between jobs meant that we were limited to one scratch buffer per
script, i.e. one scratch buffer per editing session. This is
disanalogous with quickfix lists (there can be up to 10 quickfix lists),
the other main way of displaying output utilized by the 'shell' module,
and seems like a needless limitation for some use cases (e.g., when I
exit a note, I run both 'pandoc' and 'build-index', and it would be
easiest to have a scratch buffer for each, just like I could have a
quickfix list for each, if I wanted. Note: this way of implementing
scratch buffers is still somewhat disanalogous with quickfix lists,
because you can only have one quickfix window open at a time, while
there is nothing to prevent multiple windows containing different, or
even the same, scratch buffers open at a time. But this seems A.
harmless and B. inherent to the fact that a scratch buffer is merely a
buffer, and thus even the same one could be displayed in multiple
windows, and in any event C. I would have to actually write code to
limit the user to one scratch buffer. So multiple scratch buffers are to
be allowed for now).

Eliminating 's:scratch_buf' means that there is no longer a way to save
information about scratch buffers between jobs, at least without
implementing a more complex script-local data structure. A simpler
solution is to just assume that any given job will always use a scratch
buffer of the same title, and that that title will be unique. This is a
reasonable assumption for most use cases (and also for the default
title, which is '[Job Output]' followed by the command, verbatim). So,
if no buffer exists with the title that the scratch buffer is supposed
to have, create a new one; otherwise, reuse the existing buffer.

To make this strategy possible, no longer support supplying 'title' to
'scratch_buf' in 'opts' as a list, so that we can have a different
scratch buf title depending on exit status (because then there would be
no way to know which title the existing scratch buffer is supposed to
have, unless we check for both -- which would mean both convoluted code
and probably lower performance). This is probably for the best anyway,
since that was a flourish that decreased maintainability more than it
was useful.

Also, significant bug fix: close windows containing scratch buffers when
they are open but the job produces no new output. This will depend on
whether or not the job is to print stderr, stdout, or both to the
scratch buffer, so determine whether or not there has been no output in
'shell#scratch_on_output', so that it can be determined for one, the
other, or both events. If there is no output on all desired events,
don't run the body of 'shell#scratch_on_exit', and close the window
containing the scratch buffer, if there is one.

Smaller bug fixes:
*   Use 'win_getid()' instead of 'bufwinid(bufnr('%'))', since the same
    buffer can be open in multiple windows.
*   Check for whether or not the quickfix title is *equal* to 'l:title',
    not whether or not it is matched by 'l:title'.
*   Use 'copy()' within 'map()' in 'shell#output_to_quickfix()'.
*   Make 's:title' part of 's:job_handler', so that it can use 'self'
    variable.