feat: add script for syncing all forks
Or, more precisely, merging in upstream changes to all forks. 'git'
doesn't make it easy to do this, since it doesn't allow the use of
'pull'-type commands with 'git submodule foreach'. ('git' can be forced
to, but apparently this is linked to a security vulnerability. So better
to do it manually like we do in this script). The forks are just
manually specified, so this script is meant to be updated as needed.