feat!: update to match changes in `shell` module
Greatly simplify notes configuration. Instead of calling 'mdview' and
then 'build-index' in a chain of callbacks, and collating all of the
output into one scratch buffer, simply make two asynchronous
invocations, one of Pandoc and the other of 'build-index', and let them
each print their output in their own scratch buffer.

Also simplify by, instead of doing the old 'modified' check, do what
'make' does and just compare the timestamp of the input file and the
output file to decide whether or not to run Pandoc and 'build-index' on
exit.

Delete code relating to the 'modified' check and to 'mdview', since now
we are just using the 'jobs' module directly instead of the (early in my
programming career) deprecated 'mdview' plugin.