refactor: use dictionary functions to find start
The current implementation relies on pattern matching, which is not
the only way to find the start (e.g., if we fall back on the 'default'
completer, we just use the return value of the 'default' omnifunc). So
use completer-specific functions to find the start. This allows

1.  More sophisticated checks; for example, in an 'in_context' function,
    we can decline to complete citations if we are in a YAML block.

2.  Control over the pattern used to terminate the 'while' loop that the
    citation completer uses to find the start. The 'in_context' function
    can define a different 'start_pattern' depending on e.g. the value
    of 'a:line'.

3.  No difference in implementation of the citation completer and the
    'default' completer (earlier, we had to check if the completer was
    the 'default' completer and, if so, handle it differently).

The other advantage is less branching, less reliance on 'for' loops, and
therefore improved readability.

Also FEAT: complete citations enclosed in curly braces as well
(although, at this point, don't add brace at end. Could be implemented
later).